+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Fixed rendering of type signatures concerning several identifiers. [Issue
   566](https://github.com/tweag/ormolu/issues/566).
 
+* Renamed the `--check-idempotency` flag to `--check-idempotence`.
+  Apparently only the latter is correct.
+
 ## Ormolu 0.0.5.0
 
 * Grouping of statements in `do`-blocks is now preserved. [Issue